requirements.txt
asgiref==3.2.5
astroid==2.3.3
autopep8==1.4.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.3
dateparser==0.7.2
defusedxml==0.5.0
Django==2.2
django-allauth==0.37.1
django-cors-headers==2.4.0
django-crispy-forms==1.8.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-rest-auth==0.9.3
django-role-permissions==2.2.1
django-widget-tweaks==1.4.8
django-widgets==0.1.16
djangorestframework==3.8.2
fake-factory==9999.9.9
Faker==2.0.4
gunicorn==19.7.1
idna==2.7
isort==4.3.21
jet-bridge-base==0.6.6
jet-django==0.8.9
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==2.1.0
Pillow==6.2.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
regex==2019.11.1
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.24.2
wcwidth==0.1.7
whitenoise==3.3.1
wrapt==1.11.2
